Move OpPrimitives to reth-optimism-primitives
