+++ crates/optimism/primitives/src/lib.rs
@@ -9,3 +9,22 @@
 
 pub mod bedrock;
 pub mod tx_type;
+
+pub use tx_type::OpTxType;
+
+use alloy_consensus::Header;
+use reth_node_types::NodePrimitives;
+use reth_primitives::{Block, BlockBody, Receipt, TransactionSigned};
+
+/// Optimism primitive types.
+#[derive(Debug, Default, Clone, PartialEq, Eq)]
+pub struct OpPrimitives;
+
+impl NodePrimitives for OpPrimitives {
+    type Block = Block;
+    type BlockHeader = Header;
+    type BlockBody = BlockBody;
+    type SignedTx = TransactionSigned;
+    type TxType = OpTxType;
+    type Receipt = Receipt;
+}
